👌 IMPROVE: Move format job list into jobs service
